+++ test/unit/htmlValidatorTest.js
@@ -507,6 +507,63 @@ describe('HTML Validator/Kill Validator Test', function () {
       })
     })
 
+    it('should still validate the HTML page if there is no model ', function (done) {
+      // generate the app
+      generateTestApp({
+        generateFolderStructure: true,
+        appDir: appDir,
+        viewEngine: [
+          'html: teddy'
+        ],
+        htmlValidator: {
+          enable: true,
+          separateProcess: {
+            enable: false,
+            autoKiller: false
+          },
+          showWarnings: false,
+          exceptions: {
+            requestHeader: 'partialtest'
+          }
+        },
+        onServerStart: `(app) => {process.send(app.get("params"))}`
+      }, options)
+
+      // fork the app and start it as a child process
+      const testApp = fork(path.join(appDir, 'app.js'), ['--dev'], { 'stdio': ['pipe', 'pipe', 'pipe', 'ipc'] })
+
+      // when the server starts,  check that the page has not been validated
+      testApp.on('message', (params) => {
+        request(`http://localhost:${params.port}`)
+          .get('/noModelTest')
+          .set('partialtest', 'true')
+          .expect(200, (err, res) => {
+            if (err) {
+              assert.fail(err)
+              testApp.send('stop')
+            }
+            // test the header exception in the app param is false or not there
+            let test1 = typeof res.header.partialtest === 'undefined'
+            assert.strictEqual(test1, false)
+            // check to see that the page did not validate
+            let test2 = res.text.includes('HTML did not pass validation')
+            assert.strictEqual(test2, false)
+            // kill the validator and the app
+            fkill(`:48888`, { force: true }).then(() => {
+              testApp.send('stop')
+            }, (err) => {
+              console.log(err)
+              testApp.send('stop')
+            })
+          })
+      })
+
+      // when the child process exits, finish the test
+      testApp.on('exit', () => {
+        done()
+      })
+    })
+
     it('should try to validate the HTML page because the model in the response does not hold a value that is set in the exception param', function (done) {
       // generate the app
       generateTestApp({