+++ test/unit/routesTest.js
@@ -112,6 +112,470 @@ describe('Roosevelt Routes Tests', function () {
     })
   })
 
+  it('should render a basic HTML page with a route prefix of "/prefix/"', function (done) {
+    // generate the test app
+    generateTestApp({
+      appDir: appDir,
+      generateFolderStructure: true,
+      routers: {
+        controllers: [
+          {
+            prefix: '/prefix',
+            files: ['plainHTMLController.js']
+          }
+        ]
+      },
+      onServerStart: `(app) => {process.send(app.get("params"))}`
+    }, options)
+
+    // fork and run app.js as a child process
+    const testApp = fork(path.join(appDir, 'app.js'), { 'stdio': ['pipe', 'pipe', 'pipe', 'ipc'] })
+
+    // when the app starts and sends a message back to the parent try and request the test page
+    testApp.on('message', (params) => {
+      // test the basic HTML Route
+      request(`http://localhost:${params.port}`)
+        .get('/prefix/HTMLTest')
+        .expect(200, (err, res) => {
+          if (err) {
+            testApp.send('stop')
+            assert.fail(err)
+          }
+          testApp.send('stop')
+        })
+    })
+
+    // when the child process exits, finish the test
+    testApp.on('exit', () => {
+      done()
+    })
+  })
+
+  it('should still create a router even if there\'s no leading slash in the prefix paramater', function (done) {
+    // generate the test app
+    generateTestApp({
+      appDir: appDir,
+      generateFolderStructure: true,
+      routers: {
+        controllers: [
+          {
+            prefix: 'prefix',
+            files: ['plainHTMLController.js']
+          }
+        ]
+      },
+      onServerStart: `(app) => {process.send(app.get("params"))}`
+    }, options)
+
+    // fork and run app.js as a child process
+    const testApp = fork(path.join(appDir, 'app.js'), { 'stdio': ['pipe', 'pipe', 'pipe', 'ipc'] })
+
+    // when the app starts and sends a message back to the parent try and request the test page
+    testApp.on('message', (params) => {
+      // test the basic HTML Route
+      request(`http://localhost:${params.port}`)
+        .get('/prefix/HTMLTest')
+        .expect(200, (err, res) => {
+          if (err) {
+            testApp.send('stop')
+            assert.fail(err)
+          }
+          testApp.send('stop')
+        })
+    })
+
+    // when the child process exits, finish the test
+    testApp.on('exit', () => {
+      done()
+    })
+  })
+
+  it('should render routes of all controllers within a directory using a route prefix', function (done) {
+    // generate the test app
+    generateTestApp({
+      appDir: appDir,
+      generateFolderStructure: true,
+      routers: {
+        controllers: [
+          {
+            prefix: '/prefix',
+            files: ['routerDir']
+          }
+        ]
+      },
+      onServerStart: `(app) => {process.send(app.get("params"))}`
+    }, options)
+
+    // fork and run app.js as a child process
+    const testApp = fork(path.join(appDir, 'app.js'), { 'stdio': ['pipe', 'pipe', 'pipe', 'ipc'] })
+
+    // when the app starts and sends a message back to the parent try and request the test page
+    testApp.on('message', (params) => {
+      // test the first controller in the directory
+      request(`http://localhost:${params.port}`)
+        .get('/prefix/controller1')
+        .expect(200, (err, res) => {
+          if (err) {
+            testApp.send('stop')
+            assert.fail(err)
+          }
+          // test the second controller in the directory
+          request(`http://localhost:${params.port}`)
+            .get('/prefix/controller2')
+            .expect(200, (err, res) => {
+              if (err) {
+                assert.fail(err)
+              }
+              testApp.send('stop')
+            })
+        })
+
+      // when the child process exits, finish the test
+      testApp.on('exit', () => {
+        done()
+      })
+    })
+  })
+
+  it('should render routes of controllers using a schema file within the controllers directory', function (done) {
+    // generate the test app
+    generateTestApp({
+      appDir: appDir,
+      generateFolderStructure: true,
+      routers: {
+        controllers: 'schema.js'
+      },
+      onServerStart: `(app) => {process.send(app.get("params"))}`
+    }, options)
+
+    let schema = `module.exports = [
+      {
+        prefix: '/prefix',
+        files: ['plainHTMLController.js']
+      }
+    ]`
+
+    fse.outputFileSync(path.join(appDir, 'mvc/controllers/schema.js'), schema)
+
+    // fork and run app.js as a child process
+    const testApp = fork(path.join(appDir, 'app.js'), { 'stdio': ['pipe', 'pipe', 'pipe', 'ipc'] })
+
+    // when the app starts and sends a message back to the parent try and request the test page
+    testApp.on('message', (params) => {
+      // test basic HTML Route
+      request(`http://localhost:${params.port}`)
+        .get('/prefix/HTMLTest')
+        .expect(200, (err, res) => {
+          if (err) {
+            testApp.send('stop')
+            assert.fail(err)
+          }
+          testApp.send('stop')
+        })
+
+      // when the child process exits, finish the test
+      testApp.on('exit', () => {
+        done()
+      })
+    })
+  })
+
+  // test for console output for invalid params within routers
+  let logOutputTests = [
+    {
+      logName: 'should warn that there is an invalid configuration in the "routers.controllers" parameter if one of the indices in the files array is not an object',
+      routers: {
+        controllers: [
+          'not an object',
+          {}
+        ]
+      },
+      getRequest: '/HTMLTest',
+      logMessage: 'Invalid configuration found in the "routers.controllers" parameter. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.'
+    },
+    {
+      logName: 'should warn that there is an invalid configuration in the "routers.controllers" parameter if one of the controller objects is missing the key "prefix"',
+      routers: {
+        controllers: [
+          {
+            files: ['plainHTMLController.js']
+          }
+        ]
+      },
+      getRequest: '/HTMLtest',
+      logMessage: 'Invalid configuration found in the "routers.controllers" parameter. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.'
+    },
+    {
+      logName: 'should warn that there is an invalid configuration in the "routers.controllers" parameter if the prefix includes unsafe URL Characters',
+      routers: {
+        controllers: [
+          {
+            prefix: '^invalid',
+            files: ['plainHTMLController.js']
+          }
+        ]
+      },
+      getRequest: '/HTMLTest',
+      logMessage: 'Invalid configuration found in the "routers.controllers" parameter. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.'
+    },
+    {
+      logName: 'should warn that there is an invalid configuration in the "routers.controllers" parameter if one of the controller objects is missing the key "files"',
+      routers: {
+        controllers: [
+          {
+            prefix: '/test'
+          }
+        ]
+      },
+      getRequest: '/HTMLTest',
+      logMessage: 'Invalid configuration found in the "routers.controllers" parameter. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.'
+    },
+    {
+      logName: 'should warn that there is an invalid configuration in the "routers.controllers" parameter if one of the controller objects has invalid data type for the key "files"',
+      routers: {
+        controllers: [
+          {
+            prefix: '/test',
+            files: true
+          }
+        ]
+      },
+      getRequest: '/HTMLTest',
+      logMessage: 'Invalid configuration found in the "routers.controllers" parameter. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.'
+    },
+    {
+      logName: 'should warn that there is an invalid configuration in the "routers.controllers" parameter if one of the controller objects has an empty array for the key "files"',
+      routers: {
+        controllers: [
+          {
+            prefix: '/test',
+            files: []
+          }
+        ]
+      },
+      getRequest: '/HTMLTest',
+      logMessage: 'Invalid configuration found in the "routers.controllers" parameter. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.'
+    },
+    {
+      logName: 'should warn that there is an invalid configuration in the "routers.controllers" parameter if one of the controller objects has an invalid data type within the key "files"',
+      routers: {
+        controllers: [
+          {
+            prefix: '/test',
+            files: ['plainHTMLController.js', true]
+          }
+        ]
+      },
+      getRequest: '/test/HTMLTest',
+      logMessage: 'Invalid configuration found in the "routers.controllers" parameter. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.'
+    },
+    {
+      logName: 'should warn that the app failed to load a controller file associated with a specific router if the file does not exist',
+      routers: {
+        controllers: [
+          {
+            prefix: '/prefix',
+            files: ['doesnotexist.js', 'plainHTMLController.js']
+          }
+        ]
+      },
+      getRequest: '/prefix/HTMLTest',
+      logMessage: 'failed to load the file: "doesnotexist.js" to use with the router associated with controller prefix: /prefix'
+    },
+    {
+      logName: 'should warn that the app failed to load a controllers directory associated with a specific router if the directory does not exist',
+      routers: {
+        controllers: [
+          {
+            prefix: '/prefix',
+            files: ['doesnotexist', 'plainHTMLController.js']
+          }
+        ]
+      },
+      getRequest: '/prefix/HTMLTest',
+      logMessage: 'failed to load the directory: "doesnotexist" to use with the router associated with controller prefix: /prefix'
+    },
+    {
+      logName: 'should warn that a schema file containing the controller routers could not be loaded by roosevelt',
+      routers: {
+        controllers: 'doesnotexist.js'
+      },
+      getRequest: '/HTMLTest',
+      logMessage: 'Failed to load file: "doesnotexist.js". All controllers will be routed through the app level router'
+    },
+    {
+      logName: 'should warn that there is an invalid configuration in the "routers.public" parameter if one of the indices in the files array is not an object',
+      routers: {
+        public: [
+          'not an object',
+          {}
+        ]
+      },
+      getRequest: '/css/style.css',
+      logMessage: 'Invalid configuration found in the "routers.public" parameter. Please make sure it is coded correctly. See documentation at http://github.com/rooseveltframework/roosevelt for examples.'
+    },
+    {
+      logName: 'should warn that the app failed to load a public directory associated with a specific router if the directory does not exist',
+      routers: {
+        public: [
+          {
+            prefix: '/prefix',
+            dirs: ['doesnotexist', 'css']
+          }
+        ]
+      },
+      getRequest: '/prefix/css/style.css',
+      logMessage: 'failed to load the directory: "doesnotexist" to use with the router associated with public prefix: /prefix'
+    }
+  ]
+
+  logOutputTests.forEach(test => {
+    let warnBool = false
+    let rooseveltLogOutput
+    it(test.logName, function (done) {
+      fse.outputFileSync(path.join(appDir, 'statics/css/style.css'), 'body { color: red; }')
+      // generate the test app
+      generateTestApp({
+        appDir: appDir,
+        generateFolderStructure: true,
+        routers: test.routers,
+        onServerStart: `(app) => {process.send(app.get("params"))}`
+      }, options)
+
+      // fork and run app.js as a child process
+      const testApp = fork(path.join(appDir, 'app.js'), ['--dev'], { 'stdio': ['pipe', 'pipe', 'pipe', 'ipc'] })
+
+      // when the app starts and sends a message back to the parent try and request the test page
+      testApp.on('message', (params) => {
+        request(`http://localhost:${params.port}`)
+          .get(test.getRequest)
+          .expect(200, (err, res) => {
+            if (err) {
+              testApp.send('stop')
+              assert.fail(err)
+            } else {
+              testApp.send('stop')
+            }
+          })
+      })
+
+      testApp.stderr.on('data', message => {
+        rooseveltLogOutput = message.toString()
+        if (message.toString().includes(test.logMessage)) {
+          warnBool = true
+        }
+      })
+
+      // when the child process exits, finish the test
+      testApp.on('exit', () => {
+        assert.strictEqual(warnBool, true, `Roosevelt expected the error message: ${test.logMessage} but recieved: ${rooseveltLogOutput}`)
+        done()
+      })
+    })
+  })
+
+  let redirectTests = [
+    {
+      logName: 'should concatenate a route prefix for redirection with res.redirect(address)',
+      status: 302,
+      redirectArgs: {
+        address: '/endpoint'
+      },
+      location: '/prefix/endpoint'
+    },
+    {
+      logName: 'should concatenate a route prefix for redirection with res.redirect(status, address)',
+      status: 302,
+      redirectArgs: {
+        address: '/endpoint',
+        status: '302'
+      },
+      location: '/prefix/endpoint'
+    },
+    {
+      logName: 'should not concatenate a route prefix for redirection with res.redirect(address, override)',
+      status: 302,
+      redirectArgs: {
+        address: '/endpoint',
+        override: 'true'
+      },
+      location: '/endpoint'
+    },
+    {
+      logName: 'should not concatenate a route prefix for redirection with res.redirect(status, address, override)',
+      status: 302,
+      redirectArgs: {
+        address: '/endpoint',
+        status: '302',
+        override: 'true'
+      },
+      location: '/endpoint'
+    },
+    {
+      logName: 'should log an error and redirect to \'/\' if invalid or no arguments are passed to res.redirect()',
+      status: 302,
+      redirectArgs: {},
+      location: '/',
+      error: true,
+      errorLog: 'Invalid arguments supplied to res.redirect()'
+    }
+  ]
+  redirectTests.forEach(test => {
+    it(test.logName, function (done) {
+      let location
+      let errorBool
+      fse.outputFileSync(path.join(appDir, 'statics/css/style.css'), 'body { color: red; }')
+      // generate the test app
+      generateTestApp({
+        appDir: appDir,
+        generateFolderStructure: true,
+        routers: {
+          controllers: [
+            {
+              prefix: '/prefix',
+              files: ['routerDir']
+            }
+          ]
+        },
+        onServerStart: `(app) => {process.send(app.get("params"))}`
+      }, options)
+
+      // fork and run app.js as a child process
+      const testApp = fork(path.join(appDir, 'app.js'), ['--dev'], { 'stdio': ['pipe', 'pipe', 'pipe', 'ipc'] })
+
+      // when the app starts and sends a message back to the parent try and request the test page
+      testApp.on('message', (params) => {
+        request(`http://localhost:${params.port}`)
+          .post('/prefix/redirect')
+          .send(test.redirectArgs)
+          .expect(302, (err, res) => {
+            if (err) {
+              testApp.send('stop')
+              assert.fail(err)
+            } else {
+              location = res.headers.location
+              testApp.send('stop')
+            }
+          })
+      })
+
+      if (test.error) {
+        testApp.stderr.on('data', data => {
+          if (data.includes(test.errorLog)) {
+            errorBool = true
+          }
+        })
+      }
+
+      // when the child process exits, finish the test
+      testApp.on('exit', () => {
+        assert.strictEqual(location, test.location)
+        if (test.error) assert.strictEqual(errorBool, true, 'Roosevelt was expected to print an error log if invalid or no arguments are passed to res.redirect()')
+        done()
+      })
+    })
+  })
+
   it('should use a custom port and render a basic HTML page on request', function (done) {
     // generate the test app
     generateTestApp({