+++ test/util/mvc/controllers/routerDir/controller2.js
@@ -1,0 +1,5 @@
+module.exports = (router) => {
+  router.route('/controller2').get((req, res) => {
+    res.send('test')
+  })
+}