+++ CHANGELOG.md
@@ -2,11 +2,20 @@
 
 ## Next version
 
+- If custom routers are being used, the `res.redirect()` method will now prepend the prefix to redirects that are relative to the hostname. To override this setting pass `true` as the last argument.
+- HTML validator `exceptions` param will now accept an array of values instead of just a single string value.
+- Various dependencies bumped.
+- CI improvements. 
+
+## 0.13.0
+
 - Roosevelt will now warn you if your CSS/JS compile directory is getting stale and might need to be cleaned with `npm run clean`. A new `cleanTimer` param has been added to configure or disable this check.
+- Added new param `routers` to add support for [Express Routers](https://expressjs.com/en/guide/routing.html#express-router).
 - Added new feature to create custom CSS/JS preprocessors on the fly.
 - Added environment variable to enable/disable the HTML validator autokiller.
-- Various dependencies bumped.
 - Added OpenJDK support to the Java HTML Validator.
+- Various dependencies bumped.
+- CI improvements.
 
 ## 0.12.2